requirements.txt
discord.py==0.15.1
guildwars2api==0.7.1
sqlalchemy==1.1.4